coding
Background: In the film Venom, there is a parasitic creature named Venom. People who are parasitised by it will die rapidly, although they have the appearance of regular humans. This parasite is therefore extremely dangerous to society.
Bio description: Venom assumes a liquid state when there is no host, but it enhances the host's physical and restorative abilities when parasitised, and can communicate with them telepathically. Venom is capable of picking its host and will quickly consume the host's body when it encounters an unsuitable one, causing the victim's death. It can also extend its tentacles from the host's body and deliver a fatal blow to the human.
Vulnerabilities: Fear of fire and sound waves
This is an anti-venom protection device for buildings. When a creature infested with venom comes within range of the ultrasonic distance sensor, it automatically triggers a beeping device that emits a venom-fearing sound wave to repel them, and led lights are illuminated to alert the user.